# packages/contracts/generated/openapi.v1.json
{
  "components": {
...
    "schemas": {
...
      "OtpRequestRequest": {
        "properties": {
          "locale": {
            "$ref": "#/components/schemas/Locale"
          },
          "phone": {
            "description": "Bangladesh mobile number",
            "example": "01700000001",
            "maxLength": 24,
            "minLength": 6,
            "type": "string"
          },
          "purpose": {
            "default": "LOGIN",
            "enum": [
              "LOGIN"
            ],
            "type": "string"
          }
        },
        "required": [
          "phone"
        ],
        "type": "object"
      },
      "OtpRequestResponse": {
        "properties": {
          "challengeId": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          },
          "expiresAt": {
            "example": "2026-09-18T10:49:42.206Z",
            "format": "date-time",
            "type": "string"
          },
          "hint": {
            "description": "SENT; RETRY_LATER (not delivered, request a new code later); MAY_ARRIVE (uncertain — never auto-resent)",
            "enum": [
              "SENT",
              "RETRY_LATER",
              "MAY_ARRIVE"
            ],
            "type": "string"
          }
        },
        "required": [
          "challengeId",
          "expiresAt",
          "hint"
        ],
        "type": "object"
      },
      "OtpVerifyRequest": {
        "properties": {
          "client": {
            "$ref": "#/components/schemas/ClientKind"
          },
          "code": {
            "pattern": "^\\d{6}$",
            "type": "string"
          },
          "deviceLabel": {
            "maxLength": 80,
            "type": "string"
          },
          "phone": {
            "description": "Bangladesh mobile number",
            "example": "01700000001",
            "maxLength": 24,
            "minLength": 6,
            "type": "string"
          }
        },
        "required": [
          "phone",
          "code"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "Error body for every non-2xx response (API-IMPLEMENTATION §1)",
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "details": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                },
                {
                  "type": "null"
                }
              ]
            },
            "type": "object"
          },
          "fieldErrors": {
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "message": {
            "description": "Safe, localizable message key (never PHI)",
            "type": "string"
          },
          "requestId": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          },
          "retryAfterSeconds": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "type": "object"
      },
...
      "ResponseMeta": {
        "properties": {
          "replayed": {
            "enum": [
              true
            ],
            "type": "boolean"
          },
          "requestId": {
            "example": "01a0b422-fd6f-7480-8586-444e7fc66080",
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "requestId"
        ],
        "type": "object"
      },
...
      "SessionResponse": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "accessTokenExpiresAt": {
            "example": "2026-09-18T10:49:42.206Z",
            "format": "date-time",
            "type": "string"
          },
          "csrfToken": {
            "description": "Web only; echo in X-CSRF-Token",
            "type": "string"
          },
          "refreshToken": {
            "description": "Mobile only (secure storage)",
            "type": "string"
          },
          "refreshTokenExpiresAt": {
            "example": "2026-09-18T10:49:42.206Z",
            "format": "date-time",
            "type": "string"
          },
          "user": {
            "$ref": "#/components/schemas/MeUser"
          }
        },
        "required": [
          "accessToken",
          "accessTokenExpiresAt",
          "user"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "description": "Access token (EdDSA JWT). Web keeps it in memory; mobile in secure storage.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "HMedic clinic platform API. Generated from packages/contracts (do not edit the JSON by hand). Business routes live under /api/v1; health routes at the root.",
    "title": "HMedic API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/auth/otp/request": {
      "post": {
        "operationId": "requestOtp",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66081",
              "pattern": "^[A-Za-z0-9_\\-:.]{8,191}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpRequestRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/OtpRequestResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Challenge created (identical for known/unknown phones)"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "operationId": "verifyOtp",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66081",
              "pattern": "^[A-Za-z0-9_\\-:.]{8,191}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OtpVerifyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SessionResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Session started"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/login": {
      "post": {
        "operationId": "passwordLogin",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66081",
              "pattern": "^[A-Za-z0-9_\\-:.]{8,191}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/SessionResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Session started"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/reset/complete": {
      "post": {
        "operationId": "completePasswordReset",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66081",
              "pattern": "^[A-Za-z0-9_\\-:.]{8,191}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetComplete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Done"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
    "/api/v1/auth/password/reset/request": {
      "post": {
        "operationId": "requestPasswordReset",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "example": "01a0b422-fd6f-7480-8586-444e7fc66081",
              "pattern": "^[A-Za-z0-9_\\-:.]{8,191}$",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "null"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/ResponseMeta"
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "type": "object"
                }
              }
            },
            "description": "Accepted (identical for unknown accounts)"
          },
          "400": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Validation failed"
          },
          "401": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Unprocessable (e.g. Idempotency-Key reused with another body)"
          },
          "429": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Rate limited (see Retry-After)"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "Internal error"
          }
        },
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "Local development",
      "url": "http://localhost:3000"
    }
  ],
...
}